Handle case where a tweet can't be retrieved.
When retrieving threads, it's possible to try to retrieve a tweet
the user is not authorized to see. This will return an object with
an error message instead of the usual tweet json. This change fixes
an error where an unretrieved tweet prevented display as HTML. Instead
a message is inserted explaining the tweet couldn't be retrieved.